requirements.txt
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.45.0
django-autoslug==1.9.8
django-crispy-forms==1.13.0
djongo==1.3.3
dnspython==2.1.0
dnspython==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
pymongo==3.12.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.2.4
psycopg2==2.8.6
whitenoise
django_heroku
Pillow